Added exit 0 to functions

+++ Class_Exercises/aws_cli_ec2_08052023.sh
@@ -5,15 +5,18 @@ function start_instances() {
     --filter "Name=instance-state-name,Values=stopped" \
     --query "Reservations[*].Instances[*].InstanceId" \
     --output text) )
+    
     if [ ${#stopped_instances_list[@]} -eq 0 ]; then
         echo "All instances are running! No stopped instances."
-    else
-        echo "Turning on stopped instances ids: "
-        for instance in ${stopped_instances_list[@]}; do
-            echo "Turning on instanceId: $instance"
-            aws ec2 start-instances --instance-ids $instance 
-        done
-    fi             
+        exit 0
+    fi
+    
+    echo "Turning on stopped instances ids: "
+    for instance in ${stopped_instances_list[@]}; do
+        echo "Turning on instanceId: $instance"
+        aws ec2 start-instances --instance-ids $instance 
+    done
+    echo                
 }
 
 function stop_instances() {
@@ -26,31 +29,34 @@ function stop_instances() {
     
     if [ $arr_len -eq 0 ]; then
         echo "All instances are Stopped! No running instances."
-    else
-        echo "Listing Running instances ids:"
-        for ((i = 1 ; i <= $arr_len ; i++)); do
-            echo "$i. ${running_instances_list[$i-1]}"
-        done
-        echo "$i. All"   
-        local flag=0
-        local pat="[1-$arr_len]"
-        while [ $flag -eq 0 ]
-        do
-            read -p "Which instace would you like to stop? " choice
-            if [[ $choice =~ $pat ]]; then
-                if [[ $choice -eq $arr_len+1 ]]; then
-                    echo "Stopping ALL instances!!!!"
-                    aws ec2 stop-instances --instance-ids ${running_instances_list[@]}
-                else
-                    echo "Stopping instance: ${running_instances_list[$choice-1]}"
-                    aws ec2 stop-instances --instance-ids ${running_instances_list[$choice-1]}
-                fi
-                flag=1
+        exit 0
+    fi
+    
+    echo "Listing Running instances ids:"
+    for ((i = 1 ; i <= $arr_len ; i++)); do
+        echo "$i. ${running_instances_list[$i-1]}"
+    done
+    echo "$i. All"   
+    
+    local flag=0
+    local pat="[1-$arr_len]"
+    while [ $flag -eq 0 ]
+    do
+        read -p "Which instace would you like to stop? " choice
+        if [[ $choice =~ $pat ]]; then
+            if [[ $choice -eq $arr_len+1 ]]; then
+                echo "Stopping ALL instances!!!!"
+                aws ec2 stop-instances --instance-ids ${running_instances_list[@]}
             else
-                echo "Invalid input please enter a number between 1 - $arr_len"
+                echo "Stopping instance: ${running_instances_list[$choice-1]}"
+                aws ec2 stop-instances --instance-ids ${running_instances_list[$choice-1]}
             fi
-        done                
-    fi             
+            flag=1
+        else
+            echo "Invalid input please enter a number between 1 - $arr_len"
+        fi
+    done
+    echo
 }
 
 function terminate_instance() {
@@ -62,21 +68,34 @@ function terminate_instance() {
     
     if [ $arr_len -eq 0 ]; then
         echo "No instances to show."
-    else
-        echo "Listing instances ids:"
-        for ((i = 1 ; i <= $arr_len ; i++)); do
-            echo "$i. ${instances_list[$i-1]}"
-        done
-        echo "$i. All"   
+        exit 0
+    fi
+    
+    echo "Listing instances ids:"
+    for ((i = 1 ; i <= $arr_len ; i++)); do
+        echo "$i. ${instances_list[$i-1]}"
+    done
+    echo "$i. All"   
+    
+    local flag=0
+    local pat="[1-$arr_len]"
+    while [ $flag -eq 0 ]
+    do    
         read -p "Which instace would you like to terminate? " choice
-        if [[ $choice -eq $arr_len+1 ]]; then
-            echo "Terminating ALL instances!!!!"
-            aws ec2 terminate-instances --instance-ids ${instances_list[@]}
+        if [[ $choice =~ $pat ]]; then
+            if [[ $choice -eq $arr_len+1 ]]; then
+                echo "Terminating ALL instances!!!!"
+                aws ec2 terminate-instances --instance-ids ${instances_list[@]}
+            else
+                echo "Terminating instance: ${instances_list[$choice-1]}"
+                aws ec2 terminate-instances --instance-ids ${instances_list[$choice-1]}
+            fi
+            flag=1
         else
-            echo "Terminating instance: ${instances_list[$choice-1]}"
-            aws ec2 terminate-instances --instance-ids ${instances_list[$choice-1]}
-        fi                
-    fi             
+            echo "Invalid input please enter a number between 1 - $arr_len"
+        fi
+    done
+    echo                                         
 }
 
 function create_EBS() {
@@ -84,7 +103,7 @@ function create_EBS() {
     echo "Creating EBS Volume..."
     local vol_id=$(aws ec2 create-volume \
     --volume-type gp2 \
-    --size 10 \2
+    --size 10 \
     --availability-zone us-west-2b | grep -Eo 'vol-[0-9a-zA-Z]+')
 
     local instances_list=( $(aws ec2 describe-instances \
@@ -95,15 +114,28 @@ function create_EBS() {
     
     if [ $arr_len -eq 0 ]; then
         echo "No instances to show."
-    else
-        echo "Listing instances ids:"
-        for ((i = 1 ; i <= $arr_len ; i++)); do
-            echo "$i. ${instances_list[$i-1]}"
-        done
+        exit 0
+    fi
+    
+    echo "Listing instances ids:"
+    for ((i = 1 ; i <= $arr_len ; i++)); do
+        echo "$i. ${instances_list[$i-1]}"
+    done
+    
+    local flag=0
+    local pat="[1-$arr_len]"
+    while [ $flag -eq 0 ]
+    do    
         read -p "For which instace would you like to attach the EBS volume? " choice
-        echo "Attaching EBS volume to instance: ${instances_list[$choice-1]}"
-        aws ec2 attach-volume --volume-id $vol_id --instance-id ${instances_list[$choice-1]} --device /dev/sdf
-    fi    
+        if [[ $choice =~ $pat ]]; then
+            echo "Attaching EBS volume to instance: ${instances_list[$choice-1]}"
+            aws ec2 attach-volume --volume-id $vol_id --instance-id ${instances_list[$choice-1]} --device /dev/sdf
+            flag=1
+        else
+            echo "Invalid input please enter a number between 1 - $arr_len"
+        fi
+    done
+    echo      
 }
 
 function main() {
@@ -128,7 +160,8 @@ function main() {
                 flag=1 
                 ;;
             * ) 
-                echo "Invalid Choice, Please try again!" 
+                echo "Invalid Choice, Please try again!"
+                echo 
                 ;;            
             esac
         done
